beaker: remove multi domain config resources output test
The test does not work correctly with puppet4, I hope we can re-enable
it again sometimes.
The feature works well, it's just a problem of ordering in the test
where we test an output that appears in the wrong order.

Also remove configuration that does not exist after the resource purge
and fix order of config validation on puppet4.

Change-Id: I6bffaff194cbc65432e4eab2001fd2fd8660e329