Deprecate keystone::endpoint::version
With the move to Keystone v3 API, let's deprecate version parameter and
set it to undef by default, so it will deploy Keystone versionless
endpoints.

Keep the capacity to set it to 'unset' so v2.0 will still be created, so
our users have a path of migration.

Though this patch changes the default value so from now, endpoints will
be created by default without version. If this patch breaks you, set
version to 'unset' and think about a migration to Keystone v3.

Change-Id: Ic2f741589bfbd687417e086379fc2fe60eba2024